perf(ssr): isolate open redirect detection
